requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.10.0

# Deep learning and diffusion
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
xformers>=0.0.20  # Optional, for faster attention

# Face detection and recognition
insightface>=0.7.3
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0  # Use if GPU available

# Pose detection
mediapipe>=0.10.0

# Image segmentation
# segment-anything  # Uncomment if using SAM
# git+https://github.com/facebookresearch/segment-anything.git

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.0
pyyaml>=6.0
psutil>=5.9.0  # For system monitoring
peft>=0.6.0  # For LoRA support

# Optional dependencies
# dlib>=19.24.0  # Alternative face detector (requires manual build)
# openpose-python  # Alternative pose detector (requires separate installation)
google-generativeai>=0.8.0  # Google AI Studio (Gemini) for enhanced quality assessment

# AI Image Generation APIs (optional, for API-based generation)
openai>=1.0.0  # OpenAI DALL-E API
replicate>=0.15.0  # Replicate API (optional)
requests>=2.31.0  # For API requests

